pyproject.toml
[tool.poetry]
name = "dataminingbigproject"
version = "0.1.0"
description = ""
authors = ["Rainshaw <rxg@live.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Scrapy = "^2.5.0"
scrapy-redis = "^0.7.1"
scrapy-djangoitem = "^1.1.1"
Django = "^3.2.4"
selenium = "^3.141.0"
mysqlclient = "^2.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
